+++ ripozo_sqlalchemy_tests/integration/__init__.py
@@ -3,4 +3,4 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from . import alchemymanager, columns, common, pagination, relationships
+from ripozo_sqlalchemy_tests.integration import alchemymanager, columns, common, pagination, relationships