+++ ripozo_sqlalchemy_tests/unit/alchemy_manager.py
@@ -3,6 +3,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from ripozo.exceptions import ValidationException
 from ripozo_sqlalchemy.alchemymanager import AlchemyManager, NoResultFound, NotFoundException
 
 import mock
@@ -58,3 +59,27 @@ class TestAlchemyManager(unittest2.TestCase):
         session = mock.MagicMock()
         resp = m.queryset(session)
         self.assertTrue(session.query.called)
+
+
+class TestValidateFields(unittest2.TestCase):
+    """AlchemyManager._validate_fields"""
+    def test__when_all_fields_valid__no_exception(self):
+        manager = AlchemyManager(None)
+        resp = manager._validate_fields({'key': 'value', 'key2': 'value2'}, valid_fields=['key', 'key2'])
+        self.assertIsNone(resp)
+
+    def test__when_missing_fields__validation_exception(self):
+        manager = AlchemyManager(None)
+        manager.model = mock.Mock(__name__='name')
+        self.assertRaises(ValidationException,
+                          manager._validate_fields,
+                          {'key': 'value', 'key2': 'value2'},
+                          valid_fields=['key'])
+
+    def test__when_no_valid_fields_kwarg__uses_self_fields(self):
+        class MyManager(AlchemyManager):
+            fields = ['key', 'key2']
+
+        manager = MyManager(None)
+        resp = manager._validate_fields({'key': 'value', 'key2': 'value2'})
+        self.assertIsNone(resp)